AT2Text/AI2Text frist/File kiến thức/AI2text_embeddings_patch/requirements-embeddings.txt
# === Embeddings & Vector Search ===
gensim>=4.3.2
faiss-cpu>=1.7.4
Unidecode>=1.3.7
